Initial themis_ra.py and utils.py
themis_ra.py includes functionality to parse a json config file into a dictionary,
which should be enough to close #21.
utils.py includes some basic utilities that were ripped from the krc project, and
it also houses the functionality to preprocess images.  This functionality should
probably eventually be moved to a different file, but it should be fine for now.
Closes #22.