Added initial cost function and pixel optimization.
The optimization function is set to optimize just a single pixel.
This commit does not address any batch or multiple pixel processing
at the moment.  However, given that it works for a single pixel,
we should just be able to call it in a loop once we're ready to
process an entire image.  I believe that this commit closes #36.